requirements.txt
requests
google-cloud-bigquery
pandas
db-dtypes
google-auth